inference-service/app/utils/preprocess.py
import numpy as np
from PIL import Image

IMG_SIZE = (128, 128)

def preprocess_image(image: Image.Image):
    image = image.convert("RGB")
    image = image.resize(IMG_SIZE)
    image = np.array(image) / 255.0
    image = np.expand_dims(image, axis=0)
    return image
