inference-service/requirements.txt
fastapi
uvicorn
tensorflow
numpy
pillow
python-multipart
httpx
python-dotenv
